% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_macc.R
\name{geom_macc}
\alias{geom_macc}
\title{A marginal abatement cost curve geom}
\usage{
geom_macc(fill = NULL, ...)
}
\arguments{
\item{fill}{An optional parameter to specify fill groupings to be used by `geom_macc`.}

\item{...}{Any additional arguments to pass to `geom_rect`.}
}
\description{
A `ggplot2` geom with aesthetic mappings linked to the output of `macc_prep`. Add to
a `ggplot` object to build a marginal abatement cost curve. A pre-parameterised wrapper for
`geom_rect`.
}
