% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmacc.R
\name{ggmacc}
\alias{ggmacc}
\title{Out-of-the-box marginal abatement cost curve}
\usage{
ggmacc(
  data,
  mac,
  abatement,
  fill = NULL,
  cost_threshold = NULL,
  zero_line = FALSE,
  threshold_line = FALSE,
  threshold_fade = 1
)
}
\arguments{
\item{data}{A dataframe containing marginal abatement cost and abatement data.}

\item{mac}{The variable in `data` containing marginal abatement cost data.}

\item{abatement}{The variable in `data` containing marginal abatement cost data.}

\item{fill}{An optional parameter to specify a fill variable to be used by `geom_macc`.}

\item{cost_threshold}{An optional parameter to specify a cost threshold for the marginal
abatement cost. Modifies the behaviour of `threshold_line` and `threshold_alpha`.}

\item{zero_line}{An optional parameter; logical indicating whether a solid line should be plotted
at `mac` = 0.}

\item{threshold_line}{An optional parameter; logical indicating whether to implement a horizontal
line indicating the `mac` cost threshold. Requires that `cost_threshold` be specified.}

\item{threshold_fade}{An optional parameter to set the alpha (transparency) value for plot
elements which exceed the marginal abatement cost threshold. Should be between 0 and 1; defaults
to 1 (no transparency).}
}
\description{
Wrapper compiling the functionality of the rest of the `ggmacc` package. Make a
stylish, customisable marginal abatement cost curve plot.
}
