% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macc_prep.R
\name{macc_prep}
\alias{macc_prep}
\title{Data preparation for MACC plotting}
\usage{
macc_prep(data, mac, abatement)
}
\arguments{
\item{data}{A dataframe containing marginal abatement cost and abatement data.}

\item{mac}{The name of the variable containing marginal abatement cost data.}

\item{abatement}{The name of the variable containing abatement data.}
}
\description{
Takes a dataframe containing marginal abatement cost and abatement data, and
computes bounding box variables to be passed to `geom_macc`. Pipe the results into a `ggplot()`
function and add `geom_macc`.
}
