% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_example}
\alias{small_example}
\title{Potted example MACC dataset}
\format{
A tbl_df with columns:
\describe{
 \item{mac}{A dbl indicating marginal abatement cost, arbitrary units}
 \item{abatement}{A dbl indicating abatement, arbitrary units}
 \item{cat}{A chr indicating measure category, a-e}
}
}
\source{
The author's imagination
}
\usage{
small_example
}
\description{
Potted example MACC dataset
}
\examples{
\dontrun{
 small_example
}
}
\keyword{datasets}
