% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_agroforestry}
\alias{uk_agroforestry}
\title{MACC dataset for UK silvoarable and silvopastoral systems.}
\format{
A tbl_df with columns:
\describe{
 \item{lon}{A dbl indicating grid cell longitude, decimal degrees}
 \item{lat}{A dbl indicating grid cell latitude, decimal degrees}
 \item{crop}{A chr indicating crop type}
 \item{area_ha}{A dbl indicating the area, under the specified crop in each grid cell, hectares}
 \item{co2_tyear}{A dbl indicating abatement, tonnes carbon dioxide per year}
 \item{mac_gbp_tco2}{A dbl indicating marginal abatement cost, GBP per tonne carbon dioxide}
}
}
\source{
Analysis conducted and documented at https://github.com/aj-sykes92/agroforestry-defra-prelim
}
\usage{
uk_agroforestry
}
\description{
MACC dataset for UK silvoarable and silvopastoral systems.
}
\examples{
\dontrun{
 uk_agroforestry
}
}
\keyword{datasets}
